+++ src/terra-dev-site/test/modal-manager/DisclosureComponent.jsx
@@ -13,9 +13,11 @@ import {
   getActiveDisclosureCount,
 } from '../../../disclosure-manager';
 import NavigationPrompt from '../../../navigation-prompt';
-import styles from './DisclosureComponentCommon.test.module.scss';
 import ApplicationLoadingOverlay from '../../../application-loading-overlay';
+import NotificationBannerExample from '../notification-banner/NotificationBannerExample';
 import ApplicationStatusOverlay from '../../../application-status-overlay';
+
+import styles from './DisclosureComponentCommon.test.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -156,7 +158,9 @@ class DisclosureComponent extends React.Component {
   }
 
   render() {
-    const { disclosureManager, identifier, renderHeaderAdapter } = this.props;
+    const {
+      disclosureManager, identifier, nestedIndex, renderHeaderAdapter,
+    } = this.props;
     const {
       hasPendingAction, hasLoadingOverlay, hasError, hasStatusOverlay, disclosureCount,
     } = this.state;
@@ -167,7 +171,7 @@ class DisclosureComponent extends React.Component {
 
     return (
       <ContentContainer id={identifier} className="nested-component" fill header={<h2 className={cx('header')}>Content Component</h2>}>
-        {renderHeaderAdapter ? (
+        {renderHeaderAdapter && nestedIndex > 0 ? (
           <DisclosureManagerHeaderAdapter
             title={`Disclosure - ${identifier}`}
             collapsibleMenuView={(
@@ -212,6 +216,7 @@ class DisclosureComponent extends React.Component {
         <button type="button" className="global-close-disclosure" onClick={closeMostRecentDisclosure}>{`Global Close (${disclosureCount})`}</button>
         {hasPendingAction && <NavigationPrompt description="Test Action" />}
         {hasLoadingOverlay && <ApplicationLoadingOverlay isOpen backgroundStyle="light" />}
+        <NotificationBannerExample isInitiallyClosed id={identifier} />
         {hasStatusOverlay && <ApplicationStatusOverlay message="Test status view" variant="no-data" />}
       </ContentContainer>
     );