+++ tests/jest/notification-banner/index.test.js
@@ -1,0 +1,7 @@
+import NotificationBanner from '../../../src/notification-banner';
+
+describe('notification-banner/index', () => {
+  it('should export NotificationBanner', () => {
+    expect(NotificationBanner).toBeDefined();
+  });
+});